requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# NLP & Transformers
transformers>=4.35.0
tokenizers>=0.15.0
datasets>=2.14.0
sentencepiece>=0.1.99

# Explainability (XAI)
captum>=0.6.0
shap>=0.43.0
lime>=0.2.0

# Computer Vision (for multimodal)
Pillow>=10.0.0
opencv-python>=4.8.0

# Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Adversarial Robustness
textattack>=0.3.8

# Utilities
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0

# Web/API (optional)
flask>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Indic NLP (for code-switching)
indic-nlp-library>=0.92

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
